fulcio: Added support for specifying AWS settings.
Prior to this commit, the chart did not provide a way to supply
AWS credentials for AWS KMS. This commit adds support for AWS KMS
by allowing users to supply an AWS region ID and IAM credentials.

AWS KMS users must specify the "cloudPlatform" parameter with
a value of "aws" and specify "certificateAuthority" as "kmsca".
The chart will then look for a kubernetes secret named by the
"awsCredentialsSecret" parameter.

The AWS region ID can be supplied using the newly-added
"awsRegion" parameter.

Signed-off-by: Stephen Fox <sfoxjr@equinix.com>